lab_setup/auto_scripts/pyproject.toml
[project]
name = "databricks-setup"
version = "0.1.0"
description = "Modular Databricks environment setup for Neo4j workshop"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.40.0",
    "python-dotenv>=1.0.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
]

[project.scripts]
databricks-setup = "databricks_setup.main:app"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_setup"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["databricks_setup"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
packages = ["databricks_setup"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = "databricks.*"
ignore_missing_imports = true
